refactor(langgraph): example agent graphファイル名を変更し、Gemini 2.0 Flashを使用するように更新
- `example_graph.py`を`simple_chat_graph.py`にリネームしました。
- LangchainのVertex AIモデルを`gemini-1.5-pro`から`gemini-2.0-flash`に変更しました。
- main.pyを削除しました。
